+++ spec/factories/node_route.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :node_route, class: ::Releaf::Node::Route do
+  factory :node_route, class: ::Releaf::ContentRoute do
     sequence(:node_id) {|n| n}
     sequence(:path) {|n| "path-#{n}"}
     locale "en"